server/requirements.txt
moviepy
faster_whisper
google-genai
pymongo
